+++ modules/store/payment/router/cancel_payment.js
@@ -1,30 +1,35 @@
 const { iyzipay , Iyzipay} = require("../../../../utils/iyzipay");
 const Data = require("../model")
+const ApiError = require("../../../../errors/ApiError");
 
 const route = async (req, res, next) => {
     try {
         let { userData, params } = req;
         let buyer_ip = "192.168.1.37";
         let _data = await Data.findOne({ $and:[{_id: params.id },{ author: userData.id }]}).lean();
+        if(!_data)  
+            return next(new ApiError("Payment Data not found",404));
         await iyzipay.cancel.create({
             ip: "192.168.1.37",
             paymentId: _data.paymentId,
         }, async function (err, result) {
-            console.log(result)
             if(result.status === "failure")
-                return res.status(402).send({ status: false, message: result.errorMessage, code: result.errorCode})
+                return next(new ApiError(result.errorMessage,400));
             if(result.status === "success"){
-                await Data.deleteOne({ $and: [{_id: params.id },{ author: userData.id }] })
+                let d_payment = await Data.deleteOne({ $and: [{_id: params.id },{ author: userData.id }] })
+                if(d_payment.deletedCount === 0)    
+                    return next(new ApiError("Payment delete data didn't work",400))
                 return res.status(200).send({ status: true, message: "Cancel Payment Success", data: result })
             }
         })
     } catch (error) {
-        console.log(error)
-        if(error){
-            if(error.name === "MongoError" && error.code === 11000)
-                return res.status(500).send({ status: false, message: `File Already exists!  : ${error}` })
+        if (error.name === "MongoError" && error.code === 11000) {
+          next(new ApiError(error?.message, 422));
         }
-        return res.status(500).send({ status: false, message: `Cancel Payment Error , Something Missing => ${error}`})
+        if (error.code === 27) {
+          next(new ApiError("We Don't Have Any Data", 500, null));
+        }
+        next(new ApiError(error?.message));
     }
 }
